Support login_tty from bdinfo on fsvybrid
In U-Boot, the new boot strategies set the login_tty into the kernel
command line. The kernel then copies this to the bdinfo, which shows
the value in /sys/bdinfo/login_tty and /sys/bdinfo/login_speed.

This makes setting up the login_tty more easier. We have to modify
inittab and the local environment setup in S01fssetup and S01cubea5.