Change Linux compilation procedure for localdir
F&S has added an option to allow compiling the Linux kernel from an
existing local directory, i.e. without downloading and/or extracting
the kernel from a tar file first. However when the kernel is compiled
in this separate directory, many references of different packages to
the Linux source code will fail. They expect the source code under
$(LINUX_DIR), which is output/build/linux-localir in our case.

The new procedure will copy the contents of the separate linux
directory to $(LINUX_DIR) first. But to avoid a long copy procedure
and the additional disk usage, the copy is done by hard links (if
the linux directory is on the same filesystem) or by symbolic links
(if it is on a different filesystem). This is done by the new shell
script linux/localdir_cp.

After the copy procedure is completed, a

 make distclean

is run on $(LINUX_DIR) to remove any potentially existing object
files and binaries. Otherwise the build process would overwrite the
files of the original linux directory. Then the regular build process
is started in $(LINUX_DIR) like in any other build variant.

This also requires revoking an earlier modification in the imx-lib
makefile that was necessary for our earlier implementation.