Fix libtool .la file modification style
The autotools seem to create some libtool files (with extension .la)
that refer to global paths under /usr instead of local paths in the
buildroot tree. Buildroot tries to fix this by replacing this global
/usr with the buildroot staging path (see package/pkg-autotools.mk in
define $(2)_INSTALL_STAGING_CMDS). However the sed rule to do this
also hits other paths, for example of toolchain libraries of a
toolchain installed in /usr/local/arm. This results in some bad paths
in the resulting .la files, causing compilation errors in some packages.

For example on F&S boards when trying to build the midori browser,
this causes compilation errors in libcairo and libpango as part of the
webkit package.

When looking at newer buildroot versions, there seem to be additional
issues with other libraries under /usr, too. But these issues seem to
be finally fixed in buildroot-2015.05. So this commit backports the
according pkg-autotools.mk part from buildroot-2015.05.