Replace graphic stuff with newer version to get support for 3.14
The following packages are replaced:
- freescale-imx package
- gstreamer/gst-fsl-plugins
- libfsl* packages
- x11r7/xdriver_xf86-video-imx-viv

Because buildroot is based on 3.10.53 Kernel in release 2015.08.1
we have to use an even newer version. The packages are taken from the
official buildroot git
(non release commit 0a854ce4ccc60eb58a8c53552af6b3bdc445bae9 )
Package imx-vpu uses versin 5.4.28 as it is in Yocto instead of 5.4.27
(br 2015.08.1) or 5.4.31 (br commit 0a854).
This provides a working set of packages. It shouldn't be a problem to
move to a higher version, but has not been tested yet.

config file fsimx6_std_defconfig is adjusted. There is still a problem
with xkeyboardconfig so that host-gettext has to be removed manuelly
when build process is hanging at building xkeyboard-config.