fix issues that prevent compiling (cairo, imx-lib..)
missing change in last commit
package/freescale-imx/imx-lib/imx-lib.mk is on the newer version too.

cairo needs libraries of the package xlib_libXdamage to compile, we
can't set mesa3d as dependency because this will lead to a recursive
dependency failure.

remove old Trigger Happy patch cause it is not needed anymore.