Check if /etc is writable before creating openssh keys
When booting, only try to compute any missing ssh keys if the /etc
directory is not on a read-only filesystem. Otherwise computing will
slow down the boot process considerably. And this computation is even
useless, because the keys can not be stored afterwards.

This delay is very ugly because this will happen on each boot cycle
until the script /etc/init.d/S50sshd is started at least once with a
writable /etc directory.

The test is quite simple. We try to touch /etc itself. If this is
not successful, we do not start the ssh-keygen program and /etc
remains unmodified. If the keys are available, this even speeds up
the boot process. And if there are missing keys, sshd will not
start successfully, which is the same behavior as if ssh-keygen was
started but could not write the keys.

If the touch is successful, the timestamp of /etc is updated as a
minor side effect. But this does no harm because then ssh-keygen is
started and the timestamp of /etc is updated anyway when the newly
created keys are stored into it.

There was a similar check in an earlier version of this script, but
this modification was lost when updating to buildroot-2014.08.