Switch F&S touch support from tslib to evdev
tslib has problems with multi-touch devices. There exists an extension
but this is not available in buildroot. The buildroot way seems to be
touch support based on evdev. They automatically install appropriate
evdev configs in /usr/share/X11/xorg.conf.d.

We keep the tslib way as an option. Therefore we provide a similar
configuration for tslib. Two switch from evdev to tslib, simply call
the following commands:

  cd /usr/share/X11/xorg.conf.d
  mv 10-evdev.conf 10-evdev.conf.off
  mv 10-tslib.conf.off 10.tslib.conf

To have the 10-tslib.conf.off file available, we need to provide it
in board/f+s/common/usr/share/X11/xorg.conf.d and we copy it to the
target in the board/f+s/common/final_script_std.