fix freetype changes of include paths for older packages
Freetype changes the location of the libraries between
buildroot 2015.08 and buildroot 2015.11.
(It seems that in even older versions there was another change).

Because we currently have mainly graphic stuff from newer buildroot
version (depending on 3.14.38 Kernel) and other packages from an
older verson, there are packages that search for freetype header files
at the wrong place.
We fix this by creating symlinks in the post install routine of
freetype.