Fix defconfig problems when building Linux kernel
Linux is built with the pkg-kconfig.mk file. The given defconfig file
can be enhanced with additional config fragments and finally it is
copied to .config in the linux build directory. But this may still
be a minimal config where only settings are included that differ
from the default. Therefore it is necessary to add an additional
step to bring back all undefined config symbols by actually setting
them to the default value.

In newer Linux versions, there is a make target "olddefconfig" that
does exactly this. But other Kconfig based programs like busybox and
older kernel versions like 3.0.35 do not have this target. Therefore
pkg-kconfig.mk uses target oldconfig, that should be available in all
Kconfig based programs. Target oldconfig shows all undefined symbols
on the console, and the user has to select the correct values
manually. The default value can be selected by just pressing Return.
pkg-kconfig.mk emulates this user action by using the yes command
with an empty string. This virtually presses Return after each input
prompt and thus takes the default for each request.

However strangely enough there might be differences in the resulting
.config between these two procedures (oldconfig and *_defconfig).
Some choices are handled differently. If a choice is tristate and
one of the values is given with Y, then the choice should be set to
boolean and the single value should be used. This is how *_defconfig
does it. But oldconfig does not detect the single value correctly
and assumes that the choice is undefined. Therefore it presents an
option to set it to Y or M, where M is the default. And because the
default is taken, this results in M. Then also the choice value,
which was originally set to Y, is now converted to M, which is wrong
and of course different to *_defconfig. Unfortunately we have
exactly such a defconfig file that does not work when the oldconfig
way is used.

The correct way to handle this would be to fix the bug in Kconfig in
Linux until oldconfig would behave exactly like *_defconfig. But
unfortunately this is quite complicated. Therefore this patch
modifies the build behavior of pkg-kconfig.mk.

It adds the possiblity to define $(PKG)_KCONFIG_DEFCONFIG_CMDS. By
default they are using the oldconfig variant. But the Linux Makefile
in linux.mk is now overriding this value by changing it to directly
call the appropriate *_defconfig target.

Conflicts:

	package/pkg-kconfig.mk