Make gst-fsl-plugins work on fsimx6
The gst-fsl-plugins have a generic problem in Buildroot. Buildroot
assumes that all libraries have a modified sysroot directory. Therefore
the pkgconf tool is replaced by a wrapper pkg_config that sets (among
others) the variable PKG_CONFIG_SYSROOT_DIR. This means all calls to
pkgconf will return library paths with the sysroot path prepended.
This is OK as far as the compilation and linking on the PC is affected,
but when configuration files for the target are created, the paths must
be set without the sysroot part. So distinguish between paths that need
sysroot and paths that do not.

This is done by adding an overall PKG_CONFIG_SYSROOT_DIR to the
configuration environment, but removing this part when creating the
paths for the configuration files (*.cf) by setting PKG_CONFIG_SYSROOT_DIR
to the empty string when calling pkgconf in these cases.

The last part implies a patch in package pkgconf itself. The pkg_config
wrapper must only insert the default sysroot path if the variable
PKG_CONFIG_SYSROOT_DIR is not set at all, but not if it is set to an
empty string.

Because of the old Linux kernel 3.0.35, we must also remove the uapi
part from the include directory.

We also do not need the old package/multimedia directory anymore,
gstreamer and the package gst-fsl-plugins was moved to package/gstreamer
in the meantime.