Make imx-lib work with 3.0.35 kernel
On fsimx6, we are currenly using a 3.0.35 kernel that has Vivante 3D
support at the same level than 3.10.17 kernel. Therefore we need to use
new imx-lib from 3.10.17. But the PXP device is still the old one. Add
a patch that downgrades the PXP library to work with 3.0.35 variant.

In addition, we do not have include/uapi yet, so remove it from the
imx-lib.mk file.