On F&S boards, final_script_min must create /etc/profile.d
When a minimal configuration is built and the final_script_min is
called, the script tries to copy fsalias.h to ${TARGET}/etc/profile.d.
However it does not make sure that this directory actually exists. If
not, the file will be created as a regular file named "profile.d" and any
further copy commands to this directory will fail and simply overwrite
this regular file.