Reorganize final script system for F&S boards
The common final scripts are now sourced in, not called in a new shell.
This avoids having to set variables several times.

The naming scheme is now more consitent with our regular scheme, i.e.
package_modifier-architecture, e.g. finalscript_std-fsimx6sx. This
requires renaming all final scripts and of course updateing the final
script entries in all defconfig files. While doing this, also fix some
settings in a few defconfigs that triggered legacy warnings in the
current builroot version.

Let top-level final scripts define variables like XORG_CONF or INITTAB
to define which file will be copied as xorg.conf or inittab in the
included scripts. This avoids many duplicated copy commands in the
board/architecture specific final scripts.

The finalscript_std-common now sources in the finalscript_min-common.
That means all the stuff done in the minimal configuration must not
be redone in the standard configuration, the code is re-used.

The inittab is slightly modified so that setting mdev as hotplug is
usually commented out. Only if the configuration uses mdev as device
manager, this line in inittab gets activated by removing the comment
character at the beginning of the line.

At the moment, HDMI support via v4l2 and the hdmicfg tool is only
required on fss5pv210 in standard configuration. Therefore move the
code there and remove the final script for hdmi completely.

The profile that was used in the standard configuration is a bash
profile and added way too much settings. The main reason for including
it in the past was setting the XAUTHORITY and DISPLAY variables. But
we have a profile snippet etc/profile.d/xserver.sh anyway that does
this, so we can remove the global profile completely and stay with
buildroot's standard profile. All missing aliases have been added to
etc/profile.d/fsalias.sh.

Remove pd as alias for pushd from etc/profile.d/fsalias.sh as the ash
of busybox has no pushd at all. Use color settings for the ls aliases.

Move the inittab file from board/f+s/fss5pv210 to board/f+s/common/etc
and rename it to inittab-fss5pv210. This file is used in the fss5pv210
final scripts.

Remove the unused xorg.conf in board/f+s/common.

Add own final scripts for fsimx6ul for minimal and standard config.