Add ptp-gadget package
ptp-gadget is a simple picture transfer protocol (PTP) device
implementation. PTP provides access from a USB host to a directory
of the device and allows the host to load files from and store
files to this directory (usually pictures, but basically any file
type is possible). Other than with a USB mass storage device, the
lcoal device itself can still have access to this directory while
the host is connected and any local changes to the directory contents
are immediately visible on the host.

- Add ptp-gadget.mk to download and build ptp-gadget
- Add Config.in to activate ptp-gadget and to set exported directory
- Add 0001-Fix-printf-messages.patch to fix two errors in the code
- Add S65ptp-gadget script that is copied to /etc/init.d and starts
  the PTP server at runtime

All this stuff is included in new package/ptp-gadget subdirectory.

- Include ptp-gadget in the Miscellaneous section in the generic
  package/Config.in

Remark:

You have to activate BR2_TOOLCHAIN_GLIBC_GCONV_LIBS_COPY to see the
ptp-gadget entry in menuconfig: "Toolchain" -> "Copy gconv libraries".
If you leave the next entry "Gconv libraries to copy" empty, all
gconv modules will be copied, which is quite a lot. If you do not
want this, you have to write "ISO8859-1" in this section (without
the quotation marks).