Switch NXP-based F&S boards to fs-toolchain-5.2.0-armv7ahf
The following defconfigs will use the GCC-5.2 toolchain from now on:

  fsimx6_std_defconfig
  fsimx6_min_defconfig
  fsimx6sx_std_defconfig
  fsimx6sx_min_defconfig
  fsimx6ul_std_defconfig
  fsimx6ul_min_defconfig
  fsvybrid_std_defconfig
  fsvybrid_min_defconfig
  cubea5_defconfig

In additon the floating point support is set correctly:

  Cortex-A9: NEON(/VFPv3)
  Cortex-A7: NEON/VFPv4
  Cortex-A5: NEON/VFPv4

The new toolchain fs-toolchain-5.2.0-armv7ahf does not contain a
cross-GDB, as this would also pull in ncurses and other libraries into
the toolchain that are not wanted there because then these versions
can easily clash with the versions that are built by buildroot
itself. Therefore we build a cross-GDB in Buildroot now and use the
gdbserver from there.

IMPORTANT NOTICE!
-----------------
There is a problem in GCC-4.9.x and later (including GCC-5.2) when
compiling certain X-Servers and optimizing for size at the same time.
See GCC bugtracker entry #63346. We ate F&S experience this problem at
least in buildroot-2015.11.1 and 2016.02, so switch optimization
to -O3 instead of -Os. We may set this back in the future when either
this bug is fixed or a newer X-Server does not trigger it anymore.