Fix handling of defconfig files for pkg-kconfig.mk
In package/pkg-kconfig.mk, the $(2)_KCONFIG_DEFCONFIG_CMDS was wrong,
as it concatenated the two lines to one command. By using

define $(2)_KCONFIG_DEFCONFIG_CMDS
...
endef

the linefeeds can be controlled better.

In addition in linux/linux.mk, in LINUX_KCONFIG_DEFCONFIG_CMDS, calling

  make $(BR2_LINUX_KERNEL_DEFCONFIG)_defconfig

only worked for in-tree defconfig files, if BR2_LINUX_KERNEL_USE_DEFCONFIG
was set, but not for external defconfig files anymore, when
BR2_LINUX_KERNEL_USE_CUSTOM_CONFIG was set. Fix this by copying the
defconfig file to

  $(KERNEL_ARCH_PATH)/configs/__buildroot_defconfig

first, then call

  make __buildroot_defconfig

and then remove the file again.

Remark
------
The first idea to copy the file directly to .config and then to call

  make olddefconfig

does not work for kernel 3.0.35 because target olddefconfig was added
in kernel 3.7.