Add imx-gst1-plugin to gstreamer1 package
- Add imx-gest1-plugin-4.0.9 to package/gstreamer1. This also requires
  a modification in package/gstreamer1/Config.in.

The imx-gst1-plugin package has a generic problem in Buildroot. Buildroot
assumes that all libraries have a modified sysroot directory. Therefore
the pkgconf tool is replaced by a wrapper pkg_config that sets (among
others) the variable PKG_CONFIG_SYSROOT_DIR. This means all calls to
pkgconf will return library paths with the sysroot path prepended.
This is OK as far as the compilation and linking on the PC is affected,
but when configuration files for the target are created, the paths must
be set without the sysroot part. So distinguish between paths that need
sysroot and paths that do not.

This is done by adding an overall PKG_CONFIG_SYSROOT_DIR to the
configuration environment, but removing this part when creating the
paths for the configuration files (*.cf) by setting PKG_CONFIG_SYSROOT_DIR
to the empty string when calling pkgconf in these cases.

The last part implies a patch in package pkgconf itself. The pkg_config
wrapper must only insert the default sysroot path if the variable
PKG_CONFIG_SYSROOT_DIR is not set at all, but not if it is set to an
empty string.

- Fix paths that are created in configuration files in
  output/target/usr/share/aiur_registry_1.0.arm11.cf and
  output/target/usr/share/beep_registry_1.0.arm12.cf
- This also requires a modification in package/pkgconf/pkg-config.in.