Fix installation procedure for CubeA7UL
The image for the recovery volume is the same as for the rootfs
volume. So use it twice.

When loading the images to RAM in MFG-Tool (in NBoot), the sequence
is important, because NBoot for i.MX6 behaves slightly different
than NBoot for Vybrid. The sequence is explained in the text file
install-cubea7ul.txt, so chnage this, too. While we are at it, improve
the text slightly.

The same download sequence can also work on Vybrid for CUBEA5. Also
the compilation procedure has slightly changed with newer buildroot
version. We are using an rsync based copy process for the linux
kernel now. Improve the readme-cubea5.txt file in the same spirit as
the one for CubeA7UL.