Use eudev-1.10 for fsvybrid
Linux kernels before 3.7 need userspace support from mdev or (e)udev to
load firmwares. However this support was dropped in eudev versions
after 1.10.

Therefore add support for a package eudev1 based on eudev-1.10 that
still has this firmware loading support for older kernels. As fsvybrid
is based on kernel 3.0.15, use this package by default. Also switch
cubea5_defconfig to use eudev with this package.

If we ever have fsvybrid with a newer kernel, we can drop the eudev1
package again.