Add Audio Chip configuration support
Add Audio Chip configuration support. In the file asound.state there are
the configurations for the sgtl5000 and wm9715. We have to copy this
file to the target folder that it is available in the buildroot. At the
end of the bootprocess the start script S01fssetup check if the file
asound.state exists and if the package alsactl exists. If both exists we
store the configuration of the asound.state file.