Add fsgetty for all F&S boards and modules
When no login_tty was given on the command line, for example after
issuing

  run .login_none

in U-Boot, then starting getty from the inittab failed. But as the getty
task is marked as "respawn" in inittab, it was immediately restarted,
failing over and over again. Every failure caused an error message in the
syslog, which made the syslog file grow considerably in size over time.

This patch supplements a previous modification that introduced the
wrapper script fsgetty on some customer specific F&S boards. Now
fsgetty is used on all F&S boards and modules.

fsgetty is called instead of getty in inittab. It simply calls the
regular getty if login_tty is a valid string and otherwise sleeps
indefinitely in a loop to avoid the restarting.

fsgetty can also directly read the login_speed from bdinfo. Therefore
creating the auxiliary file /tmp/login_speed in /etc/init.d/S01fssetup
is not necessary anymore. Remove this code.