Go to busybox config 1.24.2 | add fat tools to BR
Busybox is in version 1.24 in buildroot. The shipped config was 1.23.x

Renew busybox config to match new and changed parts of 1.24 version.
Besides that following tools where removed because they are activated in
buildroot: lsattr, chattr and fsck.
As a new tool fdisk was added.

In buildroot the dosfstools had been added (of course not to
min_defconfigs)