Do not use a session for matchbox
Adding an mb-applet-launcher to /usr/bin/matchbox-session in an earlier
commit triggered one new keyboard icon in the taskbar for each reboot if
the rootf was mounted read/write. This is wrong. It is caused by the
session that is written to a file. The session contains one such entry
for mb-applet-launcher after the first run. At the next run, this saved
session was restored, adding already one icon. But as there was still the
regular entry in /usr/bin/matchbox-session, a second icon was installed.
And of course now the saved session had two entries. And so on.

We solve this by using --no-session in /usr/bin/matchbox-session. Now the
session is not stored anymore and we start with a blank session each time.
However we also have to give the default apps then, because in case of
--no-session, the session is really empty.

FIXME:
The session management is not correct yet. At the moment, X is started
from /etc/init.d/S35x11, which is run by a dummy user. So the session would
be stored in /.matchbox/, which is wrong anyway. The default session should
belong to root and stored under /root/.matchbox/. And if more than one user
is available on the board, we should start a session manager who starts a
login dialog and then moves the session to the logged-in user.