Add package silex-wlan-fs for F&S boards
The Silex WLAN chip currently only has a closed source driver and also
needs specific versions of the WPA and hostapd tools. This package, which
is located in the Hardware handling section of Buildroot's menuconfig, adds
the driver and firmware files to the target. WPA tools and hostapd tools
can be selected separately, so also Buildroot's versions can be used
alternatively. However Silex does not recommend this. Like in the original
wpa_supplicant package, also a minimal wpa_supplicant.conf file is copied
to the /etc directory of the target. Without such a configuration file,
wpa_supplicant will not start.

So the provided silex-wlan-fs.mk file will copy all the requested files to
the target directory. It also adds a rule that, when building the Linux
kernel, the cfg80211.ko module will be renamed to cfg80211.ko.off, because
our package will provide an own module sxcfg80211.ko as a replacement and
having both at the same time would not work.

The Silex QCA9377 WLAN driver also needs an F&S MAC address because there
is no MAC address pre-programmed. Unfortunately the only way to set the MAC
address is by creating a file /lib/firmware/wlan/wlan_mac.bin with an entry
of the form

  Intf0MacAddress=aabbccddeeff

Of course this file has to be different on each board so it would have to
be created on-the-fly. But this is not possible on a read-only root
filesystem. However F&S U-Boot can dynamically generate an appropriate file
that will show up in Linux as /sys/bdinfo/Silex-MAC. So basically a
symbolic link pointing to this file would be sufficient. However the
firmware loading mechanism that is used by the driver to load this file can
not deal with sysfs directly because it can not determine the file size.

So this package will set up /lib/firmware/wlan/wlan_mac.bin on the target
to be a symbolic link pointing to /run/Silex-MAC and a small start script
S02silex (in /etc/init.d/ on the target) will copy the file from
/sys/bdinfo to /run early in the boot process. /run is a RAM disk (tmpfs),
so this will also work on a read-only root filesystem.

As all other buildroot packages, this package only consists of the rules of
how everything is built and installed. It does not contain the sources
themselves, or in our case the pre-built binaries. Buildroot usually
downloads the main package from the Internet. In our case, the binary
archive is part of the Buildroot based F&S release in the dl directory. It
has to be moved to Buildroot's download directory, which is usually also
called dl. There Buildroot will find and use it. If the release is
extracted with the install-sources.sh script, then the script will do this
automatically. If this does not work for some reason, for example because
the dl directory can not be located, then the archive can also be put into
the package directory istelf:

  package/silex-wlan-fs

The tar.bz2 archive contains binaries of the pre-built kernel modules,
firmware and configuration files and also the binaries of the Silex
specific versions of the WPA and hostapd tools, i.e. wpa_supplicant,
wpa_cli, wpa_passphrase, hostapd and hostapd_cli.