Add efusa7ull device tree to all fsimx6ul defconfigs
Those boards that have variants with i.MX6UL as well as i.MX6ULL need a
different device tree for each CPU variant. CUrrently this is the case
for efusA7UL. In case of i.MX6ULL, the device tree is called efusa7ull.dts.
Add this device tree to the default set of generated dtb files.