Add patches for Silex QCA9377 to bluez5_utils
F&S uses the Silex QCA9377 WLAN/Bluetooth chip on some boards. This requires
a series of patches in the bluez5 package to support this chip, i.e. provide
a command line option "qca" that loads the firmware files (rampatch and nvm)
and handle the chip accordingly.

Add the following files to package/bluez5_utils:

- 0001-QCA9377-HCI-Uart-Support.patch
- 0001-tools-bneptest.c-Remove-include-linux-if_bridge.h-to.patch
- 0002-QCA9377-UART-H4-IBS-Support-Baudrate-Fix.patch
- 0003-QCA9377-UART-Skip-NVM-BDADDR.patch
- 0004-Obex-Set-DBUS-Session-Service.patch
- 0005-Bluetoothd-Service-Enable-Compatibility-Mode.patch
- 0006-Fix-GATT-Glib-WARNING.patch
- 0007-Change-attribute-handler-registration-sequence.patch

The patches only modify the autoconf sources (Makefile.tools and some
.in files), not the files that are created by autoconf. So also add the
following modification:

- Change package/bluez5_utils/bluez5_utils.mk to run autoconf again before
  compiling the package

Remark: The patch series is originally for bluez-5.37, but it can be applied
to bluez-5.39 that is used here without any problems.