Update silex-wlan-fs package to v4.1_01272017-1
The Silex WLAN driver is closely related to the generic network stack in
Linux. So if customers change the network settings in Linux (even when they
only change non-wireless options), this often results in a non-functional
or even crashing WLAN driver when the precompiled wlan.ko module is loaded.

Therefore F&S has configured a more comprehensive set of network options
in Linux and has built a new silex-wlan-fs package with new precompiled
modules suited for this network stack. The idea is that this stack has
enough features included so that most customers can live with it and do
not need to change network settings anymore.

This commit updates Buildroot to use the new package.