Add package sterling-wlan-fs for F&S boards
The Sterling WLAN chip currently only has a backport source driver. This
package, which is located in the Hardware handling section of
Buildroot's menuconfig, adds the driver and firmware files to the
target.

So the provided sterling-wlan-fs.mk file will compile the corresponding
kernel modules and copy all the requested files to the target directory.

As all other buildroot packages, this package only consists of the rules
of how everything is built and installed. It does not contain the
sources themselves. Buildroot usually downloads the main package from
the Internet. In our case, the archive is part of the Buildroot based
F&S release in the dl directory. It has to be moved to Buildroot's
download directory, which is usually also called dl. There Buildroot
will find and use it. If the release is extracted with the install-
sources.sh script, then the script will do this automatically. If this
does not work for some reason, for example because the dl directory can
not be located, then the archive can also be put into the package
directory istelf:

      package/silex-wlan-fs

The tar.bz2 archive contains sources of the kernel backport directory to
compile the corresponding modules, firmware.