Fsimx7ulp: Remove xf86-input-tslib, set xserver to 1.17.4
tslib support in Xserver was activated but must be deactivated because
we don't use tslib. If xf86-input-tslib module is activated nevertheless
the hole xserver crashes with a segmentation fault.

Due to problems with xserver 1.20 we temporarily go to version 1.17 to
continue delivery. We'll go back to 1.20 again after more research and
stability tests.