Fix qt5webengine compilation
error:
Compilation stopped because a typedef 'snd_hwdep_info_t' was not
found/declared. This typedef is declared in files from alsalib_hwdep.
The error occured when qt5webengine was activated without the
'alsa' submodule. If qt5webengine build process would work correctly
no references to alsa should be needed.

Bringing in an compiler option with the reference to the header file at
the correct ggc command brought no difference.
QtWebEngineCore.pri has a ninja flag variable that holds two sysroot
paths. First an absolute path and later a relative path.
Both paths are correct but gcc is called in directory A with an option
to compile in directory B.
=> problem is of course second (relative) sysroot path worked in A but
not so in directory B. Because the incorrect relative path is behind
the absolute path in ninja flags only the incorrect on is considered.

Problems continue finding out where the ninja flags actually came from.
Qt5webengine uses lot's of different build tools: gn, ninja, qmake, make
and sometimes some python scripts as well.
The ninja flags with the incorrect path can be found in the
QtWebEngineCore.pri file and in a second .pri file as well.
From there the incorrect path is spreaded to many ninja.build files.
Problem is that QtWebEngineCore.pri file itself is generated.

This happens more or less with a command calling the tool gn.
With this call over 7000 files are generated from over 1300 input files.
Probably an answer will be out there in this mess.

The main problem is that the qt5webengine build process is more than
complexe. Even build files and configure like files are
generated in one or more steps using different tools interacting with each
other.

Solution:
Activating the alsa submodule changes the whole build process because
build scripts are generated differently which leads to another build
process.
With this patch we activate alsa by default for error free compilation.
At a mark of > 300MB additional rootfs size caused by webengine adding
a few KB/MB for alsa will have nearly no effect.