Use NXP version of libdrm
imx-gstreamer1 also needs an NXP version of libdrm, hosted in the IMX
sub-tree on codeaurora. (https://source.codeaurora.org).

Unfortunately the version that is used in rel_imx_4.9.88_2.0.0_ga is
based on libdrm-2.4.84, which is too old for xserver-1.20.3 that is used
here. Therfore use version rel_imx_4.14.78_1.0.0_ga. The differences in
the changes for IMX CPUs are rather minimal and it is based on
libdrm-2.4.91, which is new enough for xserver, though it is still older
than the original Buildroot version libdrm-2.4.97.