Make IMX codecs and parsers work with imx-gstreamer1
The package imx-gst1-plugin provides plugins for imx-gstreamer1 that
make the video hardware acceleration of IMX SoCs available for
gstreamer. Among other things it uses the two files

/usr/share/aiur_registry_1.0.arm11.cf
/usr/share/beep_registry_1.0.arm12.cf

to configure the demuxer/parser aiurdemux and the audio decoder beepdec.
The files list the file names and paths where the libraries for specific
mime types are located.

The libraries themselves are provided by the two packages imx-parser and
imx-codec. They are supposed to be installed under /usr/lib/imx-mm, but
buildroot does not support /etc/ld.so.conf, so there is no way to handle
libraries in different paths. All libraries have to be stored directly in
/usr/lib to be found by the loader. However this is only true for the
codecs themselves, the parsers and the audio wrappers are referenced via
the .cf files and must therefore stay in /usr/lib/imx-mm/parser and
/usr/lib/imx-mm/audio-codec/wrap respectively. So change the code in
imx-parser.mk and imx-codec.mk, that handles this behavior.

However there is a second problem. The paths that are stored in the two
.cf files above point to the buildroot directory on the host machine
where the packages are built (cross-compiled) instead of /usr/lib on the
target board itself. A solution for this is not simple and needs some
background information.

Buildroot does not only install any built packages in output/target for
the target board, but also in output/staging. Here also header files and
other intermediate files are stored that are needed for the compilation
and linking process, but are not required on the final system.

Information about all these installed files is handled by a program
called pkg-config. By calling this program, you can determine what
additional command line options are needed when you want to include one
of the installed packages in your software.

Buildroot has its own version of pkg-config in output/host/usr/bin. This
version prepends the sysroot path to all paths that are returned. So a
package that will be installed in /usr/lib on the target, will return
<sysroot>/usr/lib when calling pkg-config in the build process. This is
OK as far as the compilation and linking on the PC is affected, but not
for configuration files that are directly meant for the target. Here the
paths must be created without the sysroot part as a prefix. For our .cf
files we only need /usr/lib, not <sysroot>/usr/lib.

So we actually need to distinguish between files that need sysroot
prepended and files that do not. For this to work we need a modification
of pkg-config itself, which is part of package pkgconf. When looking at
pkg-config, that is actually only a shell script, we see that it sets
variable PKG_CONFIG_SYSROOT_DIR and this is the path that is prepended
later. By setting PKG_CONFIG_SYSROOT_DIR before calling pkg-config, a
package can provide an own path if needed, but if PKG_CONFIG_SYSROOT_DIR
is not set, this script will set it to the default sysroot path.

We change this behavior so that PKG_CONFIG_SYSROOT_DIR is only set to
the default sysroot path, if it is not set at all, but not if it is set
to an empty string. So we have three cases now:

1. PKG_CONFIG_SYSROOT_DIR already set: use this path to prepend
2. PKG_CONFIG_SYSROOT_DIR not set: use default sysroot path to prepend
3. PKG_CONFIG_SYSROOT_DIR set to empty string: do not prepend anything

This is done by simply exchanging the ":-" for a "-" in pkg-config.in.
${var:-str} uses str if var is not set or set to the empty string, but
${var-str} uses str only if var is not set at all, but keeps var if it
is set to the empty string. This is exactly what we want.

With this preparation, the real modification in package imx-gst1-plugin
is rather simple. We only need a small patch for configure.ac that sets
PKG_CONFIG_SYSROOT_DIR to the empty string when calling pkg-config for
the .cf files. That's all.

Remark:
There was a similar commit fd79de79a83bf01fa5a8d9a3f90641b629d274a4
before, but that was lost when packages were updated.

Finally only add packages imx-vpu and imx-vpuwrap to the list of
dependencies of package imx-gst1-plugin if the SoC has a VPU.