Always use (host-)gettext-gnu on F&S configurations
Buildroot-2019.05 introduces gettext-tiny for a minimal gettext setup
where Native Language Support (NLS) is not needed. However as F&S uses
a toolchain with glibc, we always have the full GNU gettext without the
need for an extra package on the target. This means we do neither activate
BR2_PACKAGE_GETTEXT_GNU nor BR2_PACKAGE_GETTEXT_TINY. But we also do not
activate BR2_SYSTEM_ENABLE_NLS. This results in a setting that uses
host-gettext-tiny on the host. Unfortunately this does not work correctly
for some packages, e.g. imx-gstreamer1.

Modify the gettext Config.in files to always use (host-)gettext-gnu.

FIXME: This is only a temporary modification. We assume that this will
get fixed in future Buildroot versions, so check again later.