Activate similar set of Qt5 modules in all F&S defconfigs
Modify the set of activated modules in the Qt5 defconfigs of F&S boards
to be as similar as possible. The modules creating large footprints like
qt53d, qt5quickcontrols, qt5quickcontrols2 and of course qt5webengine
remain switched off to have a rootfs that also fits on boards with only
128MB of NAND flash. At the moment, this set here creates a UBIFS image
with less than 100MB. If all Qt5 modules are activated, the UBIFS image
is roughly 1GB.

FIXME:
On fsimx6ul, the backend is still set to "linuxfb". Unfortunately no
OpenGLES applications are possible then. Theoretically it should be
possible to use eglfs, too, because mesa3d provides OpenGLES and EGL,
but then Qt5 complains that it does not find any DRM devices. Check
this again some day.