Add cube2.0ull device tree to cube defconfig and finalscript-cube
Those boards that have variants with i.MX6UL as well as i.MX6ULL need a
different device tree for each CPU variant. Currently this is the case
for cube2.0. In case of i.MX6ULL, the device tree is called
cube2.0ull.dts. Add this device tree to the default set of generated dtb
files.