optimization of alsa configuration file(asound.state)
For imxsgtl5000: The value for PCM Playback volume is decreased to 85
percentile of its' maximum volume(192), because at 100 % the audio
signal was bit noisy. After changing volume to 103, the audio signal is
more smooth while hearing out.

For wm9715: The capture volume is increased to 48 and some switch
controls like capture,Mono playback, headphone playback is turned
on(true)for the purpose of aplay and arecord. Unfortunately, arecord is
not working as expected. When LineIn(R/L) is applied to the port,
arecord records only half of the desired period. aplay works properly
with the changed settings in configuration file.