Replace package silex-wlan-fs with newer versions
Package silex-wlan-fs contained binaries and firmware for a closed source
version of the Silex SD-PAC module. This driver is only available for
kernel 4.1.15. With kernel 4.9.88, we switch over to an open source version
based on LEA.3.0. The package also contains firmware for Bluetooth, so
rename the package to silex-wlanbt-fs.

In fact we split the package into two parts:

silex-wlanbt-fs:
The driver itself. It is not pre-compiled anymore, it is compiled as part
of Buildroot. Even if it is strictly speaking only the WLAN driver, loading
this driver is also mandatory for Bluetooth operation, because it activates
clocks and antenna settings.

silex-wlanbt-fs-firmware:
The firmware files for WLAN and Bluetooth. Parts of these files are
specifically modified for F&S boards and modules, so do not use other
firmware files together with F&S hardware.