Do not install S50telnet by default, only if requested
Having a telnet daemon in a rootfs is often useful. But starting it
automatically at boot time is a security risk. Add a config option
for Busybox so that the startup script S50telnet is only installed
on the target if the user explicitly requests it. Up to now, it was
automatically installed if the telnetd applet was enabled in Busybox.