Add default weston.ini file
There is a config file for weston, it is called weston.ini. This file
can be used to add some configuration parameters for weston e.g. rotate
display. In this file there are different examples but all of them are
commented out so the user can manually modify this file.

Use-g2d occurs some problems e.g. display can´t be rotate or recovery
from deep sleep takes to much to redraw the image. So don´t support it
and check in next kernel/buildroot version.