Combine CubeA5 and Cube2.0 settings
CubeA5 on Vybrid with kernel 4.9.88 is now rather similar to Cube2.0 on
i.MX6UL. We can even use the same kernel binary. Extend the Cube2.0
settings (cube_defconfig, board specific files) so that they include all
the CubeA5 stuff, too. Move all board specific files to the new directory
board/f+s/cube and remove duplicate stuff. Update the readme-cube.txt.

The code is now optimized for Cortex-A5, because we assume that it is
easier for Cortex-A7 to understand code for Cortex-A5 than vice versa.

While we are at it, also remove inittab-fss5pv210, it is obsolete.