Improve XServer authority
There is a problem to start XServer applications manual as an user
because of the rights.
When the XServer get started by the Skript S40xorg it tries to create by
default the file ".Xauthority" in directory "/". This is wrong and can´t
work because of the rootfs is read-only mounted. If there is no
.Xauthority we are unable to start an application.
When the user login, the path to file ".Xauthority" is
"/var/run/.Xauthority". So the goal is to create ".Xauthority" file in
directory "/var/run". This directory is am RAM-disk.
To setup the path to "/var/run" we have to create an environment
variable called "XAUTHORITY" before we start the XServer.